requirements.txt
Flask
pymongo
flask_cors
telethon
pyJwt
flask-restful